move vendorMockData into src/data folder
