+++ mobile/src/data/vendorMockData.js
